Add files via upload
The index_logged_in.html file is an html page that will be displayed to the user after they successfully log in to infogamer. The circle-user-solid.svg file is the picture I'm using as the user's icon. In the index_logged_in.html file if the user clicks on the profile picture then the website will display a menu drop of 3 buttons that the user can navigate to.